homepage not existing on heroku fixes issues/86
