--- api/routes.js
@@ -1,43 +1,0 @@
-var riot    = require('riot');
-var Path    = require('path');
-
-
-var routes = [
-    {
-        method: 'GET',
-        path: '/{param*}',
-        handler: {
-            directory: {
-                path: Path.normalize(__dirname + '/')
-            }
-        }
-    },
-
-    {
-        method: 'GET',
-        path:'/labels',
-        handler: require("./handlers/labels_handler.js").get
-    },
-    {
-        method: 'POST',
-        path:'/labels',
-        handler: require("./handlers/labels_handler.js").create
-    },
-    {
-        method: 'PUT',
-        path:'/labels',
-        handler: require("./handlers/labels_handler.js").update
-    },
-    {
-        method: 'DELETE',
-        path:'/labels/{name}',
-        handler: require("./handlers/labels_handler.js").remove
-    },
-    {
-        method: 'GET',
-        path:'/login', 
-        handler: require('./handlers/authentication_handler.js')
-    }
-];
-
-module.exports = routes;
