+++ handlers/labels_handler.js
