+++ lib/db_handlers.js
