+++ lib/parse_issues.js
