+++ test/gh_to_db_parser.test.js
@@ -1,5 +1,5 @@
 var test             = require("tape");
-var parseIssues      = require("../api/parse_issues.js");
+var parseIssues      = require("../lib/parse_issues.js");
 var issueArray       = require("../example_issues.json");
 var parsedIssues     = require("./fixtures/gh_to_db_parser_fixtures");
 var parsedIssueArray = [parsedIssues.parsedIssue1, parsedIssues.parsedIssue2];
