+++ test/runner.js
@@ -1,6 +1,6 @@
-// 'use strict';
-//
-// require('./db_handlers.test')
-// require('./authentication_handler.test.js');
-// require('./get_issues.test.js');
-// require('./labels_handler.test.js');
+'use strict';
+
+require('./db_handlers.test')
+require('./authentication_handler.test.js');
+require('./get_issues.test.js');
+require('./labels_handler.test.js');
