+++ views/layout_header.html
@@ -2,7 +2,7 @@
     <head>
         <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
         <title>GitHub Tudo List!</title>
-        <link rel="icon" href="https://www.dwyl.io/images/favicon.ico">
+        <link rel="shortcut icon" href="https://www.dwyl.io/images/favicon.ico" type="image/x-icon">
         <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/materialize/0.97.0/css/materialize.min.css">
         <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
     </head>
